refactor(ingest): one source of truth for the extraction contract (#164)
The extraction contract existed as two hand-maintained rewrites — the
initial prompt and the re-extract prompt — plus a third stale copy in
dead code. Editing one and forgetting the other silently diverged the
paths, and it had already happened: the two disagreed on the `line_type`
vocabulary, on the arithmetic invariant, and on whether the ADDITIVE
pricing branch exists at all.

The shared contract now lives in four modules that BOTH builders
interpolate, extending the pattern `brand-icon-prompt.ts` already used:

  prompt-contract.ts      versions, model, agent hygiene, psql discipline,
                          date self-check, OCR audit, active lessons
  document-read-prompt.ts Phase 0 — how to read a PDF / image / .eml
  line-item-prompt.ts     item schema, line_type vocabulary, two-level
                          modifier rule, pricing logic, worked examples
  items-sql.ts            the parameterized SQL fragments, killing 8
                          copy sites

prompt.ts 1875 -> 1225 lines; reextract-prompt.ts 458 -> 325. The
*rendered* re-extract prompt goes the other way, 458 -> 1226 lines,
because it was missing most of the contract: it had no document-read
phase, no worked examples, no date self-check, and an item schema whose
SQL required 11 fields its prose never named. Measured consequence of
that gap: re-extract dropped `product_id` on 28.5% of product lines
(283/396) versus 0.2% at ingest, and `tax_minor` on 94.4%.

Drift reconciled deliberately, not merged blindly. The notable calls:

  line_type       open vocabulary, `fee` dropped as a recommended value
                  (it overlapped `surcharge`/`service_fee` and existed
                  only because the two prompts disagreed). No DB enum,
                  no CHECK — production already holds 18 values.
  arithmetic      re-extract's wording; prompt.ts's was self-
                  contradictory about whether tax rows are items
  INCLUSIVE       prompt.ts's both-branch text. re-extract defined only
                  INCLUSIVE, so its one directive about parent prices
                  was "REDUCE" — the direct cause of #165
  items JSON      `$items$` dollar-quoting at all 8 sites; single quotes
                  hard-fail on `McDonald's` and silently roll back
  versions        one PROMPT_VERSION = "3.0". REEXTRACT_PROMPT_VERSION
                  is deleted — nothing read it, and two number lines in
                  one column had "1.8" sorting newer than "2.19"
  metadata        re-extract now appends the prior extraction block to
                  `extraction_history` instead of destroying it

Dead code removed: `processReceipt` (targeted the pre-v1 `receipts`
tables, zero importers), `askClaude`, and `getSessionJsonlPath` (a
duplicate of the one in `langfuse.ts` that callers actually import).

No behavior change is intended in this commit. No new extraction rules —
those are #165/#166, which now get written once.

Verified: tsc clean; both prompts rendered to disk and diffed against
the originals — 489/511 distinctive instruction lines survive verbatim,
and all 22 differences are either a deliberate decision above or a
String.raw escaping artifact (converting away from String.raw is part of
this change: it was leaking literal backslashes into the agent's window).
No `undefined`, no `[object Object]`, no unresolved interpolation.
`$items$` present at 6 + 5 sites, single-quoted items JSON at none.

Closes #164
